Module-call dispatch splits out of calls.rs and the embedded http fixture's designed wall joins the baseline
